refactor(ui): extract AdviseMixin for modal message-box helpers
Move the eight Qt QMessageBox-based advisory/question helpers out of
L5RMain into l5r/ui/advise.py:

* advise_successfull_import
* advise_error
* advise_warning
* ask_warning
* ask_to_save
* ask_to_upgrade
* not_enough_xp_advise
* warn_about_missing_books

L5RMain now inherits AdviseMixin alongside L5RCMCore, so every
existing self.advise_* / self.ask_* call site keeps working unchanged.
warn_about_missing_books's "api.character.books" reference (which
came in via "from l5r.l5rcmcore import *" in main.py) becomes an
explicit "import l5r.api.character.books" in the new module.

main.py: 2503 -> 2406 lines.
