+++ l5r/main.py
@@ -46,6 +46,7 @@ import l5r.widgets as widgets
 import l5r.dialogs as dialogs
 
 from l5r.l5rcmcore import *
+from l5r.ui.advise import AdviseMixin
 from l5r.ui.helpers import (
     new_horiz_line,
     new_item_groupbox,
@@ -57,7 +58,7 @@ from l5r.util import log
 from l5r.util.settings import L5RCMSettings
 
 
-class L5RMain(L5RCMCore):
+class L5RMain(AdviseMixin, L5RCMCore):
 
     default_size = QtCore.QSize(820, 720)
     default_point_size = 8.25
@@ -2132,91 +2133,6 @@ class L5RMain(L5RCMCore):
             self.wounds[i][1].setText(str(i_stacked))
             self.wounds[i][2].setText(str(i_stacked_wounds) if i_stacked_wounds else '')
 
-    def advise_successfull_import(self, count):
-        settings = L5RCMSettings()
-        if not settings.app.advise_successful_import:
-            return
-        msgBox = QtWidgets.QMessageBox(self)
-        msgBox.setWindowTitle('L5R: CM')
-        msgBox.setText(
-            self.tr("{0} data pack(s) imported succesfully.").format(count))
-        do_not_prompt_again = QtWidgets.QCheckBox(
-            self.tr("Do not prompt again"), msgBox)
-        # PREVENT MSGBOX TO CLOSE ON CLICK
-        do_not_prompt_again.blockSignals(True)
-        msgBox.addButton(QtWidgets.QMessageBox.Ok)
-        msgBox.addButton(do_not_prompt_again, QtWidgets.QMessageBox.ActionRole)
-        msgBox.setDefaultButton(QtWidgets.QMessageBox.Ok)
-        msgBox.setIcon(QtWidgets.QMessageBox.Information)
-        msgBox.exec_()
-        if do_not_prompt_again.checkState() == QtCore.Qt.Checked:
-            settings.app.advise_successful_import = False
-
-    def advise_error(self, message, dtl=None):
-        msgBox = QtWidgets.QMessageBox(self)
-        msgBox.setWindowTitle('L5R: CM')
-        msgBox.setTextFormat(QtCore.Qt.RichText)
-        msgBox.setText(message)
-        if dtl:
-            msgBox.setInformativeText(dtl)
-        msgBox.setIcon(QtWidgets.QMessageBox.Critical)
-        msgBox.setDefaultButton(QtWidgets.QMessageBox.Ok)
-        msgBox.exec_()
-
-    def advise_warning(self, message, dtl=None):
-        msgBox = QtWidgets.QMessageBox(self)
-        msgBox.setTextFormat(QtCore.Qt.RichText)
-        msgBox.setWindowTitle('L5R: CM')
-        msgBox.setText(message)
-        if dtl:
-            msgBox.setInformativeText(dtl)
-        msgBox.setIcon(QtWidgets.QMessageBox.Warning)
-        msgBox.setDefaultButton(QtWidgets.QMessageBox.Ok)
-        msgBox.exec_()
-
-    def ask_warning(self, message, dtl=None):
-        msgBox = QtWidgets.QMessageBox(self)
-        msgBox.setTextFormat(QtCore.Qt.RichText)
-        msgBox.setWindowTitle('L5R: CM')
-        msgBox.setText(message)
-        if dtl:
-            msgBox.setInformativeText(dtl)
-        msgBox.setIcon(QtWidgets.QMessageBox.Warning)
-        msgBox.addButton(QtWidgets.QMessageBox.Ok)
-        msgBox.addButton(QtWidgets.QMessageBox.Cancel)
-        msgBox.setDefaultButton(QtWidgets.QMessageBox.Cancel)
-        return msgBox.exec_() == QtWidgets.QMessageBox.Ok
-
-    def ask_to_save(self):
-        msgBox = QtWidgets.QMessageBox(self)
-        msgBox.setWindowTitle('L5R: CM')
-        msgBox.setText(self.tr("The character has been modified."))
-        msgBox.setInformativeText(self.tr("Do you want to save your changes?"))
-        msgBox.addButton(QtWidgets.QMessageBox.Save)
-        msgBox.addButton(QtWidgets.QMessageBox.Discard)
-        msgBox.addButton(QtWidgets.QMessageBox.Cancel)
-        msgBox.setDefaultButton(QtWidgets.QMessageBox.Save)
-        return msgBox.exec_()
-
-    def ask_to_upgrade(self, target_version):
-        msgBox = QtWidgets.QMessageBox(self)
-        msgBox.setWindowTitle('L5R: CM')
-        msgBox.setText(
-            self.tr("L5R: CM v%s is available for download.") % target_version)
-        msgBox.setInformativeText(
-            self.tr("Do you want to open the download page?"))
-        msgBox.addButton(QtWidgets.QMessageBox.Yes)
-        msgBox.addButton(QtWidgets.QMessageBox.No)
-        msgBox.setDefaultButton(QtWidgets.QMessageBox.No)
-        return msgBox.exec_()
-
-    def not_enough_xp_advise(self, parent=None):
-        if parent is None:
-            parent = self
-        QtWidgets.QMessageBox.warning(parent, self.tr("Not enough XP"),
-                                  self.tr("Cannot purchase.\nYou've reached the XP Limit."))
-        return
-
     def closeEvent(self, ev):
         # update interface last time, to set unsaved states
         self.update_from_model()
@@ -2363,19 +2279,6 @@ class L5RMain(L5RCMCore):
 
     def get_health_rank(self, idx):
         return self.wounds[idx][1].text()
-
-    def warn_about_missing_books(self):
-
-        text = self.tr("<h3>Missing books</h3>")
-        text += self.tr("<p>To load this character you need this additional books:</p>")
-        dtl_text = u"<ul>"
-        for b in api.character.books.get_missing_dependencies():
-            dtl_text += "<li>{book_nm} &gt;= {book_ver}</li>".format(
-                book_nm=b.name, book_ver=b.version)
-        dtl_text += u"</ul>"
-
-        self.advise_error(text, dtl_text)
-
 
 # MAIN ###
